[#COD-40][BE] feat : mypage / problem

+++ backend/src/main/java/com/kw/repository/CategoryRepository.java
@@ -4,6 +4,8 @@ import com.kw.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer>, QuerydslPredicateExecutor<Category> {
-
+public interface CategoryRepository extends JpaRepository<Category, Long>, QuerydslPredicateExecutor<Category> {
+	
+	
+	
 }
